Remove props
