+++ src/screens/TaskScreen.js
@@ -50,49 +50,49 @@ const TaskScreen = () => {
     return;
   };
 
-  const removeTask = (key) => {
-    let newTasks = [...tasks].filter((task) => task.key !== key);
-    setTasks(newTasks);
-  };
+  // const removeTask = (key) => {
+  //   let newTasks = [...tasks].filter((task) => task.key !== key);
+  //   setTasks(newTasks);
+  // };
 
-  const editTask = (taskId, newTitle) => {
-    setTasks((prev) =>
-      prev.map((task) => (task.id === taskId ? newTitle : task))
-    );
-  };
+  // const editTask = (taskId, newTitle) => {
+  //   setTasks((prev) =>
+  //     prev.map((task) => (task.id === taskId ? newTitle : task))
+  //   );
+  // };
 
-  const switchStar = (key) => {
-    let newTasks = tasks.map((task) => {
-      if (task.key === key) {
-        task.star = !task.star;
-      }
-      return task;
-    });
-    setTasks(newTasks);
-  };
+  // const switchStar = (key) => {
+  //   let newTasks = tasks.map((task) => {
+  //     if (task.key === key) {
+  //       task.star = !task.star;
+  //     }
+  //     return task;
+  //   });
+  //   setTasks(newTasks);
+  // };
 
-  const inSearchBarKeyDown = (e) => {
-    console.log(e);
+  // const inSearchBarKeyDown = (e) => {
+  //   console.log(e);
 
-    if (e.key === "Alt") {
-      if (window.location.pathname === "/search") {
-        inboxRef.current.click();
-      }
-      console.log("focus on search in inSearchBarKeyDown");
-      inputRef.current.focus();
-    } else if (e.key === "ArrowUp") {
-      console.log("arrow up(input down!)");
-      archiveRef.current.click();
-    } else if (e.key === "ArrowDown") {
-      console.log("Arrow down(inSearchBarKeyDown)");
-      inputRef.current.focus();
-      inboxRef.current.click();
-    }
-  };
+  //   if (e.key === "Alt") {
+  //     if (window.location.pathname === "/search") {
+  //       inboxRef.current.click();
+  //     }
+  //     console.log("focus on search in inSearchBarKeyDown");
+  //     inputRef.current.focus();
+  //   } else if (e.key === "ArrowUp") {
+  //     console.log("arrow up(input down!)");
+  //     archiveRef.current.click();
+  //   } else if (e.key === "ArrowDown") {
+  //     console.log("Arrow down(inSearchBarKeyDown)");
+  //     inputRef.current.focus();
+  //     inboxRef.current.click();
+  //   }
+  // };
 
-  const handleSearchBarClick = () => {
-    searchRef.current.click();
-  };
+  // const handleSearchBarClick = () => {
+  //   searchRef.current.click();
+  // };
 
   const filteredTasks = tasks.filter((task) =>
     task.title.toLowerCase().includes(search.toLowerCase())
@@ -118,15 +118,7 @@ const TaskScreen = () => {
           </SideScreenContainer>
           <MainScreenContainer>
             <SectionTasksContainer>
-              <SectionTasks
-                tasks={filteredTasks}
-                sectionType={sectionType}
-                completeTask={completeTask}
-                removeTask={removeTask}
-                editTask={editTask}
-                switchStar={switchStar}
-                todayDate={todayDate}
-              />
+              <SectionTasks tasks={filteredTasks} />
             </SectionTasksContainer>
 
             <TaskInputContainer>
